requirements.txt
anyio==4.8.0
attrs==25.3.0
blinker==1.9.0
boto3==1.36.24
botocore==1.36.24
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.13.1
Flask==3.1.0
fsspec==2024.6.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsii==1.110.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
publication==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
